requirements.txt
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.10.7
future==0.18.3
idna==3.4
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai-whisper==20230314
packaging==23.0
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.2.0
pytube==12.1.3
regex==2023.3.23
requests==2.28.2
six==1.16.0
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
